refactor(orzma_vt): split the interpreter tests into interpreter/tests/*
`interpreter.rs` had grown to 2388 lines, 1660 of which were one
`mod tests` holding 114 tests. Move them into 21 modules under
`interpreter/tests/`, following the layout `screen.rs` already uses:
the parent file keeps `#[cfg(test)] mod tests;`, `interpreter/tests.rs`
holds the imports, the shared fixtures, and the module declarations, and
each child opens with `use super::*;`.

The axis is one module per control function, not per notation or per
dispatch entry point. The seven-bit, raw C1, and UTF-8 spellings of a
sequence reach the same handler, so they belong together; conversely
`DECSET`/`DECRST` fans out to origin mode, mouse, keypad, and the
alternate screen, which are separate contracts and separate modules.
The `?47`, `?1047`, and `?1049` tests stay in one file because four of
them exercise how `?47` and `?1049` interact over the cursor checkpoint
and would be orphaned by a split along mode numbers.

This is a pure move: no test body or doc comment changed, and the suite
reports the same 545 passed / 4 failed, the four being the RED tests
that `apc_dispatch` has yet to satisfy.
